Fix the Debye model GibbsFreeEnergyTask
The Debye model Gibbs energy task would fail because the force constants
were being read from the database documents, but they are not calculated
in VASP when the Debye model is used. The temperature and Gibbs free
energies returned from the `get_debye_model_gibbs` function were changed
to be returned as lists rather than NumPy arrays because the NumPy arrays
are not properly converted to lists when using `json.dumps`.

The `tools.analysis.get_debye_model_gibbs` function was also changed
because the number of elements needs to be `(t_max-t_min)/t_step +1` in
order to have enough elements for a closed interval from
`t_min` to `t_max`.