Fix phonon calculation analysis
Previous changes moved the `force_constants` processing into another loop
for phonon calculations to separate the processing from the Debye model.
This broke the phonon post-processing because the `force_constants` were
being appended to a list of outputs instead of applied to the ones that
already existed. Moving the `force_constants` processing back into the
loop fixed this.