Write the Gibbs energy data to a `gibbs` database collection
GibbsFreeEnergy task is renamed to GibbsFreeEnergyToDbTask to conform
with other similar FireTasks. Data about the relaxed structure's formula and
the EOS type and QHA type are added to the DB entry. Does not check for the
presence of a `db_file` as in other FireTasks because the current
implementation requires a DB (data is not pulled from the `fw_spec`).