Use a more sensible default for Gibbs deformations
The deformation arrays go from np.identity(3)*(0.9) to
np.identity(3)*1.1. These seem like they are intended to go from -10% vol to
+10% vol, but really go from -27.1% volume to +33.1% volume because (0.9^3=0.729).
The change here reflects that. 21 deformations is too many to be necessary for a
good EOS fitting and is a waste of computing hours. It should still be an odd
number to include the equilibirum. 11 is more appropriate.

See http://dx.doi.org/10.1016/j.commatsci.2009.12.006 for more

t_step was changed to a more reasonable 10 K. There is so much walltime compared
to calculation time, it doesn't make sense to have such a large t_step.